// frontend/src/api/generated/model/paginatedResponseTask.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gantry Board API
 * AI Agent Orchestration Kanban Board API
 * OpenAPI spec version: 0.1.0
 */
import type { PaginatedResponseTaskDataItem } from './paginatedResponseTaskDataItem';

export interface PaginatedResponseTask {
  data: PaginatedResponseTaskDataItem[];
  limit: number;
  offset: number;
  total: number;
}